Add files via upload
In this project ,financial data comprising of historical share price and quarterly revenue reportings from various online sources was extracted using y-finance , a Python library as well as  webscraping with BeautifulSoup on Tesla and Gamestop stocks. After collecting this data i used plotly to  visualize it in a dashboard and identify patterns or trends over the years.